backend-python/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
openai
numpy
scikit-learn
pandas
python-multipart